feat : 회원 카카오페이 기능 추가

+++ BE/src/main/java/com/ssafy/tati/util/Amount.java
@@ -7,12 +7,12 @@ import lombok.ToString;
 @ToString
 @Getter @Setter
 public class Amount {
-    private Long total;         //전체 결제 금액
-    private Long tax_free;      //비과세 금액
-    private Long vat;           //부과세 금액
-    private Long point;         //사용한 포인트 금액
-    private Long discount;      //할인금액
-    private Long green_deposit; //컵보증금
+    private int total;         //전체 결제 금액
+    private int tax_free;      //비과세 금액
+    private int vat;           //부과세 금액
+    private int point;         //사용한 포인트 금액
+    private int discount;      //할인금액
+    private int green_deposit; //컵보증금
 }
